Added a section about long term transitional artifacts.
Basically, I am suggesting that if your module is something that is likely to evolve, then you / we should consider producing separate artifacts using classifiers, so both versions can live on in parallel.  Not sure we _really_ want this, but if it seems likely, it is wise to at least mention it here.  I could produce an example doing this as well, if you agree it is a path we want to recommend.

Otherwise, feel free to roll back this change.